backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic-settings==2.5.2
psycopg[binary]==3.2.13
psycopg-pool>=3.2.0
PyJWT[crypto]==2.9.0
cryptography>=42.0.0
